+++ tests/commands-anx-options/assist-options.test.ts
@@ -16,6 +16,7 @@ describe('AssistOptions', () => {
     [{ '--cleanup': null }, ['--cleanup']],
     [{ '--content': null }, ['--content']],
     [{ '--content-of': 'A' }, ['--content-of=A']],
+    [{ '--content-of': ['A', 'B'] }, ['--content-of=A', '--content-of=B']],
     [{ '--jobs': 2 }, ['--jobs=2']],
     [{ '--jobs': '3' }, ['--jobs=3']],
     [{ '--message': 'A' }, ['--message=A']],