+++ tests/commands-anx-options/push-anx-options.test.ts
@@ -14,6 +14,7 @@ describe('PushAnxOptions', () => {
     [{ '--cleanup': null }, ['--cleanup']],
     [{ '--content': null }, ['--content']],
     [{ '--content-of': 'A' }, ['--content-of=A']],
+    [{ '--content-of': ['A', 'B'] }, ['--content-of=A', '--content-of=B']],
     [{ '--jobs': 2 }, ['--jobs=2']],
     [{ '--jobs': '3' }, ['--jobs=3']],
     [{ '--no-content': null }, ['--no-content']],