+++ tests/commands-anx-options/satisfy-options.test.ts
@@ -1,6 +1,6 @@
 import * as anx from '../../src/index.ts';
 
-describe('DropunusedOptions', () => {
+describe('SatisfyOptions', () => {
   let repositoryPath: string;
   let myAnx: anx.GitAnnexAPI;
 
@@ -9,16 +9,16 @@ describe('DropunusedOptions', () => {
     myAnx = anx.createAccessor(repositoryPath);
   });
 
-  const tests: [anx.DropunusedOptions, string[]][] = [
-    [{ '--from': 'A' }, ['--from=A']],
+  const tests: [anx.SatisfyOptions, string[]][] = [
+    [{ '--all': null }, ['--all']],
+    [{ '--content-of': 'A' }, ['--content-of=A']],
+    [{ '--content-of': ['A', 'B'] }, ['--content-of=A', '--content-of=B']],
     [{ '--jobs': 2 }, ['--jobs=2']],
     [{ '--jobs': '3' }, ['--jobs=3']],
-    [{ '--json': null }, ['--json']],
-    [{ '--json-error-messages': null }, ['--json-error-messages']],
   ];
 
-  test.each(tests)('DropunusedOptions "%o"', async (anxOptions, expected) => {
-    const rslt = await myAnx.dropunused('all', anxOptions, { noOp: true });
+  test.each(tests)('SatisfyOptions "%o"', async (anxOptions, expected) => {
+    const rslt = await myAnx.satisfy(undefined, anxOptions, { noOp: true });
     expect(rslt.exitCode).toBeNaN();
     expect(rslt.args).toEqual(expect.arrayContaining(expected));
   });